Add the DISABLE_INCOMING_REPORTS flag

+++ src/index.js
@@ -13,7 +13,7 @@ import { parsePeers } from './peer.js';
 printAsciiArt();
 
 async function main() {
-    const { keyPair, peerSizeLimitMB: envFilePeerSizeLimitMB } = await parseEnvFile();
+    const { keyPair, peerSizeLimitMB: envFilePeerSizeLimitMB, disableIncomingReports } = await parseEnvFile();
     const peersDirectoryPath = path.join('hinter-core-data', 'peers');
     console.log('Parsing peers...');
     const peerSizeLimitMB = envFilePeerSizeLimitMB ?? 1024;
@@ -28,11 +28,13 @@ async function main() {
     }, 60000);
 
     console.log('Preparing to connect...');
-    // Create 2 Corestore instances per peer in a local directory
+    // Create Corestore instances per peer in a local directory
     await Promise.all(peers.map(async (peer) => {
-        const incomingCorestore = new Corestore(path.join('hinter-core-data', '.storage', peer.publicKey, 'incoming'));
-        await incomingCorestore.ready();
-        peer.incomingCorestore = incomingCorestore;
+        if (!disableIncomingReports) {
+            const incomingCorestore = new Corestore(path.join('hinter-core-data', '.storage', peer.publicKey, 'incoming'));
+            await incomingCorestore.ready();
+            peer.incomingCorestore = incomingCorestore;
+        }
 
         const outgoingCorestore = new Corestore(path.join('hinter-core-data', '.storage', peer.publicKey, 'outgoing'));
         await outgoingCorestore.ready();
@@ -47,14 +49,16 @@ async function main() {
     swarm.on('connection', (conn, peerInfo) => {
         const peer = peers.find(peer => peer.publicKey === Buffer.from(peerInfo.publicKey).toString('hex'));
         if (peer) {
-            const incomingStream = peer.incomingCorestore.replicate(conn);
-            incomingStream.on('error', (err) => {
-                if (err.message.includes('conflict detected')) {
-                    console.log(`Conflict detected with ${peer.alias}. Deleting incoming storage and exiting to allow restart.`);
-                    fs.rmSync(path.join('hinter-core-data', '.storage', peer.publicKey, 'incoming'), { recursive: true, force: true });
-                    process.exit(0);
-                }
-            });
+            if (!disableIncomingReports) {
+                const incomingStream = peer.incomingCorestore.replicate(conn);
+                incomingStream.on('error', (err) => {
+                    if (err.message.includes('conflict detected')) {
+                        console.log(`Conflict detected with ${peer.alias}. Deleting incoming storage and exiting to allow restart.`);
+                        fs.rmSync(path.join('hinter-core-data', '.storage', peer.publicKey, 'incoming'), { recursive: true, force: true });
+                        process.exit(0);
+                    }
+                });
+            }
             peer.outgoingCorestore.replicate(conn);
             peer.connection = conn;
             console.log(`Connected to ${peer.alias}!`);
@@ -64,14 +68,16 @@ async function main() {
     });
 
     await Promise.all(peers.map(async (peer) => {
-        peer.incomingLocaldrive = new Localdrive(path.join(peersDirectoryPath, `${peer.alias}-${peer.publicKey}`, 'incoming'));
+        if (!disableIncomingReports) {
+            peer.incomingLocaldrive = new Localdrive(path.join(peersDirectoryPath, `${peer.alias}-${peer.publicKey}`, 'incoming'));
 
-        const incomingHyperdriveKeyPair = crypto.keyPair(crypto.data(b4a.concat([b4a.from(peer.publicKey, 'hex'), keyPair.publicKey])));
-        peer.incomingHyperdrive = new Hyperdrive(peer.incomingCorestore, incomingHyperdriveKeyPair.publicKey);
-        await peer.incomingHyperdrive.ready();
+            const incomingHyperdriveKeyPair = crypto.keyPair(crypto.data(b4a.concat([b4a.from(peer.publicKey, 'hex'), keyPair.publicKey])));
+            peer.incomingHyperdrive = new Hyperdrive(peer.incomingCorestore, incomingHyperdriveKeyPair.publicKey);
+            await peer.incomingHyperdrive.ready();
 
-        peer.incomingDiscovery = swarm.join(peer.incomingHyperdrive.discoveryKey, { client: true, server: false });
-        await peer.incomingDiscovery.flushed();
+            peer.incomingDiscovery = swarm.join(peer.incomingHyperdrive.discoveryKey, { client: true, server: false });
+            await peer.incomingDiscovery.flushed();
+        }
 
         peer.outgoingLocaldrive = new Localdrive(path.join(peersDirectoryPath, `${peer.alias}-${peer.publicKey}`, 'outgoing'));
 
@@ -88,22 +94,26 @@ async function main() {
 
     await Promise.all(peers.map(async (peer) => {
         // Do the initial mirror
-        const initialIncomingMirror = peer.incomingHyperdrive.mirror(peer.incomingLocaldrive);
-        await initialIncomingMirror.done();
-        console.log(`${peer.alias} initial incoming: ${JSON.stringify(initialIncomingMirror.count)}`);
+        if (!disableIncomingReports) {
+            const initialIncomingMirror = peer.incomingHyperdrive.mirror(peer.incomingLocaldrive);
+            await initialIncomingMirror.done();
+            console.log(`${peer.alias} initial incoming: ${JSON.stringify(initialIncomingMirror.count)}`);
+        }
 
         const initialOutgoingMirror = peer.outgoingLocaldrive.mirror(peer.outgoingHyperdrive);
         await initialOutgoingMirror.done();
         console.log(`${peer.alias} initial outgoing: ${JSON.stringify(initialOutgoingMirror.count)}`);
 
         // Mirror detected incoming changes in hyperdrive
-        (async () => {
-            for await (const { } of peer.incomingHyperdrive.watch()) {
-                const incomingMirror = peer.incomingHyperdrive.mirror(peer.incomingLocaldrive);
-                await incomingMirror.done();
-                console.log(`${peer.alias} detected incoming: ${JSON.stringify(incomingMirror.count)}`);
-            }
-        })();
+        if (!disableIncomingReports) {
+            (async () => {
+                for await (const { } of peer.incomingHyperdrive.watch()) {
+                    const incomingMirror = peer.incomingHyperdrive.mirror(peer.incomingLocaldrive);
+                    await incomingMirror.done();
+                    console.log(`${peer.alias} detected incoming: ${JSON.stringify(incomingMirror.count)}`);
+                }
+            })();
+        }
 
         // Mirror detected outgoing changes in localdrive
         fs.watch(path.join(peersDirectoryPath, `${peer.alias}-${peer.publicKey}`, 'outgoing'), { recursive: true }, async () => {
